styled the form
