+++ friends/src/components/FriendsList/FriendForm/FriendForm.css
@@ -5,6 +5,7 @@
 .input-form {
   width: 300px;
   padding: 10px 20px;
+  font-size: 18px;
 }
 
 .form-container {
@@ -15,9 +16,12 @@
 }
 .label {
   padding-right: 20px;
+  font-size: 20px;
 }
 
 .btn {
-  padding: 8px 10px;
-  margin-top: 10px;
+  padding: 8px 20px;
+  margin-top: 20px;
+  font-size: 18px;
+  background-color: #ddd;
 }
